+++ src/Shared/LayoutRenderers/AspNetMvcControllerLayoutRenderer.cs
@@ -20,11 +20,11 @@ namespace NLog.Web.LayoutRenderers
     /// </remarks>
     /// <example>
     /// <code lang="NLog Layout Renderer">
-    /// ${aspnet-mvc-controller}    
+    /// ${aspnet-mvc-controller}
     /// </code>
     /// </example>
     [LayoutRenderer("aspnet-mvc-controller")]
-    public class AspNetMvcControllerRenderer : AspNetMvcLayoutRendererBase
+    public class AspNetMvcControllerLayoutRenderer : AspNetMvcLayoutRendererBase
     {
         /// <summary>
         /// Renders the specified ASP.NET Application variable and appends it to the specified <see cref="StringBuilder" />.
