+++ tests/Shared/LayoutRenderers/AspNetRequestUrlRendererTests.cs
@@ -21,7 +21,7 @@ using System.IO;
 
 namespace NLog.Web.Tests.LayoutRenderers
 {
-    public class AspNetRequestUrlRendererTests : LayoutRenderersTestBase<AspNetRequestUrlRenderer>
+    public class AspNetRequestUrlRendererTests : LayoutRenderersTestBase<AspNetRequestUrlLayoutRenderer>
     {
         [Fact]
         public void UrlPresentRenderNonEmpty_Default()
@@ -175,7 +175,7 @@ namespace NLog.Web.Tests.LayoutRenderers
         }
 #endif
 
-        private static AspNetRequestUrlRenderer CreateRenderer(string hostBase, string queryString = "", string scheme = "http", string page = "/", string pathBase = "", string rawTarget = null)
+        private static AspNetRequestUrlLayoutRenderer CreateRenderer(string hostBase, string queryString = "", string scheme = "http", string page = "/", string pathBase = "", string rawTarget = null)
         {
             var (renderer, httpContext) = CreateWithHttpContext();
